requirements.txt
# === FRAMEWORK PRINCIPAL ===
chainlit==2.5.5

# === CACHE ET RATE LIMITING (PRIORITÉ 1) ===
redis>=5.0.0,<6.0.0
fastapi-cache2[redis]>=0.2.0,<1.0.0
fastapi>=0.104.0,<1.0.0
uvicorn[standard]>=0.24.0,<1.0.0

# === COMMUNICATION HTTP ===
httpx>=0.27.0,<0.29.0
aiohttp>=3.9.0,<4.0.0
requests>=2.31.0,<3.0.0

# === TRAITEMENT DE DONNÉES ===
pandas>=2.1.0,<3.0.0
numpy>=1.24.0,<2.0.0
pillow>=10.0.0,<11.0.0

# === TRAITEMENT AUDIO (PRIORITÉ 3) ===
librosa>=0.10.0,<1.0.0
soundfile>=0.12.0,<1.0.0
pydub>=0.25.0,<1.0.0

# === TRAITEMENT IMAGES (PRIORITÉ 4) ===
opencv-python>=4.8.0,<5.0.0
python-multipart>=0.0.6,<1.0.0

# === UTILITAIRES CORE ===
python-dotenv>=1.0.0,<2.0.0
pydantic>=2.4.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0
loguru>=0.7.0,<1.0.0

# === SÉCURITÉ ===
cryptography>=41.0.0,<45.0.0
passlib>=1.7.4,<2.0.0

# === TESTS ===
pytest>=7.4.0,<8.0.0
pytest-asyncio>=0.21.0,<1.0.0
